+++ src/KeyVault/KeyVault/Models/PSKeyVaultRoleDefinition.cs
@@ -1,4 +1,4 @@
-﻿using Azure.Security.KeyVault.Administration.Models;
+﻿using Azure.Security.KeyVault.Administration;
 using System.Linq;
 
 namespace Microsoft.Azure.Commands.KeyVault.Models
@@ -33,11 +33,11 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
         {
             Id = roleDefinition.Id;
             Name = roleDefinition.Name;
-            Type = roleDefinition.Type;
+            Type = roleDefinition.Type?.ToString();
             RoleName = roleDefinition.RoleName;
             Description = roleDefinition.Description;
-            RoleType = roleDefinition.RoleType;
-            AssignableScopes = roleDefinition.AssignableScopes.ToArray();
+            RoleType = roleDefinition.RoleType?.ToString();
+            AssignableScopes = roleDefinition.AssignableScopes.Select(s => s.ToString()).ToArray();
             Permissions = roleDefinition.Permissions.Select(permission => new PSKeyVaultPermission(permission)).ToArray();
         }
     }